Moved client app out of dotnet app, moved AppCode to ContextRepository to keep it separate from future AppCode, made INewNestRepository generic
