--- NewNest/ClientApp/.eslintrc.json
@@ -1,7 +1,0 @@
-{
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-        "ecmaVersion": 6,
-        "sourceType": "module"
-    }
-}
