+++ NewNest/ContextRepository/INewNestContext.cs
