+++ NewNestClient/.eslintrc.json
@@ -1,0 +1,7 @@
+{
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {
+        "ecmaVersion": 6,
+        "sourceType": "module"
+    }
+}
