+++ NewNestClient/package.json
@@ -1,0 +1,58 @@
+{
+  "name": "newnest",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "bootstrap": "^4.3.1",
+    "connected-react-router": "6.5.2",
+    "history": "4.10.1",
+    "jquery": "^3.4.1",
+    "merge": "1.2.1",
+    "popper.js": "^1.16.0",
+    "react": "16.11.0",
+    "react-dom": "16.11.0",
+    "react-redux": "7.1.1",
+    "react-router": "5.1.2",
+    "react-router-dom": "5.1.2",
+    "react-scripts": "^3.2.0",
+    "reactstrap": "8.1.1",
+    "redux": "4.0.4",
+    "redux-thunk": "2.3.0",
+    "svgo": "1.3.0"
+  },
+  "devDependencies": {
+    "@types/history": "4.7.3",
+    "@types/jest": "24.0.19",
+    "@types/node": "12.11.6",
+    "@types/react": "16.9.9",
+    "@types/react-dom": "16.9.2",
+    "@types/react-redux": "7.1.5",
+    "@types/react-router": "5.1.2",
+    "@types/react-router-dom": "5.1.0",
+    "@types/reactstrap": "8.0.6",
+    "@typescript-eslint/parser": "^2.5.0",
+    "cross-env": "6.0.3",
+    "eslint": "^6.5.1",
+    "eslint-plugin-flowtype": "^3.13.0",
+    "eslint-plugin-import": "2.18.2",
+    "eslint-plugin-jsx-a11y": "6.2.3",
+    "eslint-plugin-react": "7.16.0",
+    "typescript": "3.6.4"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "cross-env CI=true react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
+    "lint": "eslint ./src/**/*.ts ./src/**/*.tsx"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
+}
