+++ NewNestClient/src/components/Layout.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import { Container } from 'reactstrap';
+import NavMenu from './NavMenu';
+
+export default (props: { children?: React.ReactNode }) => (
+    <React.Fragment>
+        <NavMenu/>
+        <Container>
+            {props.children}
+        </Container>
+    </React.Fragment>
+);
